+++ locales/es/strings.vb
@@ -30,6 +30,11 @@
     Public Const REPLACEFIELDCOUNT0 As String = "Su base de datos puede estar desactualizada. Favor proceda a actualizarla para obtener los campos de la tabla "
     Public Const REPLACEFOREIGNKEYINSERT As String = "No se encuentran todos los registros necesarios en la tabla $4"
 
+    'Transfer Data
+    Public Const ERRORCONNECTDESTINYSERVER As String = "Error al conectar al servidor destino: "
+    Public Const ERRORCONNECTSOURCESERVER As String = "Error al conectar al servidor destino: "
+    Public Const ERRORDIFFERENTSCHEMAS As String = "El esquema de la tabla fuente no es el mismo que el de la tabla destino"
+
     'Windows Services
     Public Const ERRORSERVICENOTFOUND As String = "Servicio no encontrado; asegurese de que este instalado: {0}"
     Public Const ERRORSTARTINGSERVICE As String = "Error al iniciar el servicio: {0}"
@@ -40,4 +45,28 @@
 
     'GDI
     Public Const GDIUNKNOWNFORMAT As String = "Formato desconocido: "
+
+    'Dates
+    Public Const JAN As String = "Enero"
+    Public Const FEB As String = "Febrero"
+    Public Const MAR As String = "Marzo"
+    Public Const APR As String = "Abril"
+    Public Const MAY As String = "Mayo"
+    Public Const JUN As String = "Junio"
+    Public Const JUL As String = "Julio"
+    Public Const AUG As String = "Agosto"
+    Public Const SEP As String = "Septiembre"
+    Public Const OCT As String = "Octubre"
+    Public Const NOV As String = "Noviembre"
+    Public Const DEC As String = "Diciembre"
+
+    'User Errors
+    Public Const ERRORFORMATTELEPHONE As String = "El teléfono no esta en el formato correcto"
+    Public Const ERRORPERMISIONAREAMANAGER As String = "Solo un gerente de área puede habiltar este permiso"
+    Public Const ERRORPERMISIONENTRY As String = "Su usuario no dispone de los permisos necesarios para entrar al aplicativo"
+    Public Const ERRORPERMISIONOPTION As String = "Su usuario no dispone de los permisos necesarios para esta opción"
+    Public Const ERRORPERMISIONBRANCH As String = "Esta sucursal no puede ejecutar esta opción"
+
+    'Configuration Constants
+    Public Const MAILESCAPE As String = "@." 'Represents no mail given
 End Module